+++ CHANGELOG.md
@@ -13,6 +13,7 @@ project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Add switch docker container network command
 - Add edit or show configuration of docker-compose.yml command
 - Update command docs
+- Build publish script
 
 
 ## [1.0.12] - 2022-11-10